CMake: Fix boost dependencies when consuming. (#11)
### Why 
CMake: When importing the library, you would manually have to download
and link the different boost libraries used here.

```
[cmake] CMake Error at build/vcpkg_installed/x64-linux-release/share/SimpleAmqpClient/SimpleAmqpClientTargets.cmake:60 (set_target_properties):
[cmake]   The link interface of target "SimpleAmqpClient::SimpleAmqpClient" contains:
[cmake] 
[cmake]     Boost::chrono
[cmake] 
[cmake]   but the target was not found.  Possible reasons include:
[cmake] 
[cmake]     * There is a typo in the target name.
[cmake]     * A find_package call is missing for an IMPORTED target.
[cmake]     * An ALIAS target is missing.
```

I suspect that is because the install-config.cmake file did not include
the components.
Updating the vcpkg baseline also allow us to consume the header-only
components of boosts in the same way as other libraries.

Furthermore, we can save a string copy per field in BasicMessage by
accepting by-value and moving in. Will require change on the outside to
get the effect of it though.

### What was changed
* Updated vcpkg baseline.
* Updated required CMake version.
* Added Boost components to install-config.cmake.
* Updated the way we include/link the boost components (due to new vcpkg
baseline).
* Changed some `const std::string&` parameters to `std::string` with
`std::move`.